+++ jsonargparse/cli.py
@@ -2,9 +2,9 @@
 
 import inspect
 from typing import Any, Callable, Dict, List, Optional, Type, Union
-from .actions import ActionConfigFile
+from .actions import ActionConfigFile, _ActionPrintConfig, remove_actions
 from .core import ArgumentParser
-from .deprecated import cli_return_parser_message, deprecation_warning
+from .deprecated import deprecation_warning_cli_return_parser
 from .optionals import get_doc_short_description
 from .util import default_config_option_help
 
@@ -18,14 +18,13 @@ def CLI(
     config_help: str = default_config_option_help,
     set_defaults: Optional[Dict[str, Any]] = None,
     as_positional: bool = True,
-    return_parser: bool = False,
     **kwargs
 ):
-    """Function for simple creation of command line interfaces.
+    """Simple creation of command line interfaces.
 
     Creates an argument parser from one or more functions/classes, parses
     arguments and runs one of the functions or class methods depending on what
-    was parsed. If the components argument is not given, then the components
+    was parsed. If the 'components' parameter is not given, then the components
     will be all the locals in the context and defined in the same module as from
     where CLI is called.
 
@@ -35,12 +34,12 @@ def CLI(
         config_help: Help string for config file option in help.
         set_defaults: Dictionary of values to override components defaults.
         as_positional: Whether to add required parameters as positional arguments.
-        return_parser: ``[DEPRECATED]`` Whether to return the parser instead of parsing and running.
         **kwargs: Used to instantiate :class:`.ArgumentParser`.
 
     Returns:
         The value returned by the executed function or class method.
     """
+    return_parser = kwargs.pop('return_parser', False)
     caller = inspect.stack()[1][0]
     if 'description' not in kwargs:
         kwargs['description'] = caller.f_globals.get('__doc__')
@@ -67,7 +66,7 @@ def CLI(
         if set_defaults is not None:
             parser.set_defaults(set_defaults)
         if return_parser:
-            deprecation_warning((CLI, 'return_parser'), cli_return_parser_message)
+            deprecation_warning_cli_return_parser()
             return parser
         cfg = parser.parse_args(args)
         cfg_init = parser.instantiate_classes(cfg)
@@ -76,14 +75,18 @@ def CLI(
     subcommands = parser.add_subcommands(required=True)
     comp_dict = {c.__name__: c for c in components}
     for name, component in comp_dict.items():
-        subparser = ArgumentParser()
+        description = get_help_str(component, parser.logger)
+        subparser = ArgumentParser(description=description)
         subparser.add_argument('--config', action=ActionConfigFile, help=config_help)
         subcommands.add_subcommand(name, subparser, help=get_help_str(component, parser.logger))
-        _add_component_to_parser(component, subparser, as_positional, config_help)
+        added_args = _add_component_to_parser(component, subparser, as_positional, config_help)
+        if not added_args:
+            remove_actions(subparser, (ActionConfigFile, _ActionPrintConfig))
 
     if set_defaults is not None:
         parser.set_defaults(set_defaults)
     if return_parser:
+        deprecation_warning_cli_return_parser()
         return parser
     cfg = parser.parse_args(args)
     cfg_init = parser.instantiate_classes(cfg)
@@ -102,15 +105,22 @@ def get_help_str(component, logger):
 def _add_component_to_parser(component, parser, as_positional, config_help):
     kwargs = {'as_positional': as_positional, 'sub_configs': True}
     if inspect.isfunction(component):
-        parser.add_function_arguments(component, **kwargs)
+        added_args = parser.add_function_arguments(component, as_group=False, **kwargs)
+        if not parser.description:
+            parser.description = get_help_str(component, parser.logger)
     else:
-        parser.add_class_arguments(component, **kwargs)
+        added_args = parser.add_class_arguments(component, **kwargs)
         subcommands = parser.add_subcommands(required=True)
         for key in [k for k, v in inspect.getmembers(component) if callable(v) and k[0] != '_']:
-            subparser = ArgumentParser()
+            description = get_help_str(getattr(component, key), parser.logger)
+            subparser = ArgumentParser(description=description)
             subparser.add_argument('--config', action=ActionConfigFile, help=config_help)
-            subparser.add_method_arguments(component, key, **kwargs)
+            added_subargs = subparser.add_method_arguments(component, key, as_group=False, **kwargs)
+            added_args += [f'{key}.{a}' for a in added_subargs]
+            if not added_subargs:
+                remove_actions(subparser, (ActionConfigFile, _ActionPrintConfig))
             subcommands.add_subcommand(key, subparser, help=get_help_str(getattr(component, key), parser.logger))
+    return added_args
 
 
 def _run_component(component, cfg):